PDF_QUESTIONS_ANSWERING/requirements.txt
fastapi
uvicorn
python-multipart
python-dotenv
streamlit
requests
langchain
langchain-community
cohere
PyPDF2
faiss-cpu
